v0.4.0-beta: Theme overhaul, Media Library upgrade, AI Persona system, Artist Intel, AuxDeck dual-bus routing
Theme System:
- Delete dark.qss, Enterprise Pro default, ThemePalette central color registry
- Strip 200+ hardcoded setStyleSheet calls across 45+ files
- Module dock framing with steel-blue borders, 12px font minimum

Media Library (ported from Mcaster1AMP):
- FTS5 full-text search, 7 preset categories, album art cache
- AI Intel badges, category sidebar, scan-into-category
- Right-click: Rating, AutoDJ Weight, Assign to Category, Play in AUX Deck

AI Integration (35% complete):
- 6 providers (Ollama, Claude, ChatGPT, Grok, Gemini, Venice)
- 15 preset AI personas (Radio DJ, Podcast, Sports, TV, Church, Producer)
- Artist Intel Dialog: 8 research buttons + Discovery, per-tab save/load
- AI DJ Agent Browser, Playlist Generator Pro, Daypart Scheduler Pro
- PersonaManager with 3-tier resolution (daypart > surface > global)

DeckPlayer:
- Crossfader top-aligned, Playlist/AutoDJ permanently embedded
- AutoDJ auto-plays on enable, S-curve crossfade fix
- Loop boundary guard, ring buffer overflow fix, rich ICY metadata

AuxDeck v3.0:
- Full DeckPlayer-class: seek, hot cues, loop, pitch/speed, BPM
- Dual-bus audio routing: AIR + CUE independent PortAudio streams
- AIR/CUE toggle buttons, device collision prevention
- Track History + Recordings tabs, WAV recording capability

Infrastructure:
- Portable app: all storage in app directory, STANDARDS.md
- Code signing: CMake + NSIS + installer
- Sub-surface tab switching fix
- Preferences: min/max/resize, screen-aware sizing
- All 5 SQL dialects synced (9 tables)
- 117 files changed, ~10,000 lines

+++ Modules/DeckModule/StreamReader.h
@@ -71,8 +71,8 @@ private:
     static constexpr int kRingFrames = 1 << 18;  // 262144 frames ≈ 5.5 s @ 48 kHz
     static constexpr int kRingMask   = kRingFrames - 1;
     std::vector<float> m_ring;                     // size = kRingFrames * 2
-    std::atomic<int>   m_ringW{0};
-    std::atomic<int>   m_ringR{0};
+    std::atomic<int64_t> m_ringW{0};
+    std::atomic<int64_t> m_ringR{0};
 
     void ringWrite(const float* stereoData, int frames);
 };
